start

+++ module10/HelloWorld.cl
@@ -34,7 +34,7 @@ __kernel void mod_cl(__global const float *a,
     int gid = get_global_id(0);
 
     if(b[gid]!=0)
-        result[gid] = a[gid] % b[gid];
+        result[gid] = (int)a[gid] % (int)b[gid];
     else
         result[gid] = -9999;
 }
